changed affine type to float

+++ lib/common/include/type_aliases.hpp
@@ -11,7 +11,7 @@ using V4 = Eigen::Vector4f;
 using M3 = Eigen::Matrix<float, 3, 3, Eigen::RowMajor>;
 using M4 = Eigen::Matrix<float, 4, 4, Eigen::RowMajor>;
 
-using A3 = Eigen::Affine3d;
+using A3 = Eigen::Affine3f;
 
 // rotation transformations
 using Q = Eigen::Quaternionf;
